feat : Payment 원장에 취소 데이터 저장 (cancelAmount, cancelReason, canceledAt)
- Payment insert-only 원장 설계 확립: CONFIRM/CANCEL/DEPOSIT 타입별 이벤트 기록
- Toss cancels 배열에서 lastTransactionKey 기준으로 현재 취소 항목 추출 (get(0) 버그 수정)
- PaymentsResponse.currentCancelEntry() 메서드로 중복 추출 로직 통합
- cancelAmount 타입 Long → int 통일 (PaymentsCancels DTO, Payment 엔티티)
- CancelEntryNotFoundException 추가 (취소 항목 미존재 시 fail-fast)
- V7: cancel 필드 3개 추가, V8: cancel_amount BIGINT → INT 변경

+++ Mavis-Infrastructure/src/main/java/com/mavis/infrastructure/outer/api/tosspayments/dto/PaymentsCancels.java
@@ -3,7 +3,7 @@ package com.mavis.infrastructure.outer.api.tosspayments.dto;
 import java.time.ZonedDateTime;
 
 public record PaymentsCancels(
-        Long cancelAmount,
+        int cancelAmount,
         String cancelReason,
         Long taxFreeAmount,
         Long taxExceptionAmount,
